+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - add missing name attributes to classmethods
 - fixed circular imports
+- Prevent Surface2D triangulation crash.
 
 ### Refactor
 -